+++ Server/Models/Monsters/TreeMonster.cs
@@ -27,7 +27,7 @@ namespace Server.Models.Monsters
         {
         }
 
-        public override int Attacked(MapObject ob, long power, Element element, bool canReflect = true, bool ignoreShield = false, bool canCrit = true, bool canStruck = true)
+        public override long Attacked(MapObject ob, long power, Element element, bool canReflect = true, bool ignoreShield = false, bool canCrit = true, bool canStruck = true)
         {
             return base.Attacked(ob, 1, element, ignoreShield, canCrit);
         }
